+++ trunk/sqltoy-orm/src/main/java/org/sagacity/sqltoy/plugins/nosql/ElasticSearchPlugin.java
@@ -21,10 +21,10 @@ import org.sagacity.sqltoy.executor.QueryExecutor;
 import org.sagacity.sqltoy.model.DataSetResult;
 import org.sagacity.sqltoy.model.PaginationModel;
 import org.sagacity.sqltoy.utils.BeanUtil;
-import org.sagacity.sqltoy.utils.CommonUtils;
 import org.sagacity.sqltoy.utils.HttpClientUtils;
 import org.sagacity.sqltoy.utils.MongoElasticUtils;
 import org.sagacity.sqltoy.utils.ResultUtils;
+import org.sagacity.sqltoy.utils.StringUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -191,24 +191,8 @@ public class ElasticSearchPlugin {
 		// 不支持指定查询集合的行列转换
 		ResultUtils.calculate(sqlToyConfig, resultSet, null);
 		// 将结果数据映射到具体对象类型中
-		resultSet.setRows(
-				ResultUtils.wrapQueryResult(resultSet.getRows(), wrapFields(resultSet.getLabelNames()), resultClass));
+		resultSet.setRows(ResultUtils.wrapQueryResult(resultSet.getRows(),
+				StringUtil.humpFieldNames(resultSet.getLabelNames()), resultClass));
 		return resultSet;
-	}
-
-	private static String[] wrapFields(String[] fields) {
-		if (fields == null)
-			return null;
-		String[] aliasFields = new String[fields.length];
-		System.arraycopy(fields, 0, aliasFields, 0, fields.length);
-		int aliasIndex = 0;
-		for (int i = 0; i < aliasFields.length; i++) {
-			aliasIndex = aliasFields[i].indexOf(":");
-			if (aliasIndex != -1) {
-				aliasFields[i] = aliasFields[i].substring(aliasIndex + 1).trim();
-			}
-		}
-		String[] aliasNames = CommonUtils.humpFieldNames(aliasFields);
-		return aliasNames;
 	}
 }
