+++ trunk/sqltoy-orm/src/main/java/org/sagacity/sqltoy/utils/MacroIfLogic.java
@@ -3,116 +3,18 @@
  */
 package org.sagacity.sqltoy.utils;
 
-import java.io.Closeable;
-import java.io.IOException;
 import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 /**
  * @project sagacity-sqltoy4.0
- * @description 提供一些常用的函數處理
+ * @description 提供针对sql中 @if(:paramName1>=value1 && :paramName2!=value2)
+ *              性质的逻辑判断,返回true或false,适用于sql和mongo等所有查询语句中使用
  * @author chenrenfei <a href="mailto:zhongxuchen@gmail.com">联系作者</a>
- * @version id:CommonUtils.java,Revision:v1.0,Date:2017年12月9日
+ * @version id:MacroIfLogic.java,Revision:v1.0,Date:2017年12月9日
  */
-public class CommonUtils {
-	/**
-	 * 定义日志
-	 */
-	private final static Logger logger = LoggerFactory.getLogger(CommonUtils.class);
-
-//	/**
-//	 * @TODO 读取文件存为字符串
-//	 * @param file
-//	 * @param charset
-//	 * @return
-//	 */
-//	public static String readFileAsString(Object file, String charset) {
-//		return inputStream2String(getFileInputStream(file), charset);
-//	}
-//
-//	/**
-//	 * @todo 获得指定路径的文件
-//	 * @param file 文件路径like:classpath:xxx.xml或xxx.xml
-//	 * @return
-//	 */
-//	public static InputStream getFileInputStream(Object file) {
-//		if (file == null)
-//			return null;
-//		try {
-//			if (file instanceof InputStream)
-//				return (InputStream) file;
-//			if (file instanceof File)
-//				return new FileInputStream((File) file);
-//
-//			String realFile = (String) file;
-//			if (StringUtil.isBlank(realFile))
-//				return null;
-//			// 文件路径
-//			if (new File(realFile).exists()) {
-//				return new FileInputStream(realFile);
-//			}
-//			if (StringUtil.indexOfIgnoreCase(realFile.trim(), "classpath:") == 0)
-//				realFile = realFile.trim().substring(10).trim();
-//			if (realFile.charAt(0) == '/')
-//				realFile = realFile.substring(1);
-//			InputStream result = Thread.currentThread().getContextClassLoader().getResourceAsStream(realFile);
-//			if (result == null) {
-//				try {
-//					Enumeration<URL> urls = Thread.currentThread().getContextClassLoader().getResources(realFile);
-//					URL url;
-//					while (urls.hasMoreElements()) {
-//						url = urls.nextElement();
-//						result = new FileInputStream(url.getFile());
-//						if (result != null)
-//							break;
-//					}
-//				} catch (Exception e) {
-//					e.printStackTrace();
-//				}
-//			}
-//			return result;
-//		} catch (FileNotFoundException fn) {
-//			fn.printStackTrace();
-//		}
-//		return null;
-//	}
-//
-//	/**
-//	 * 转换InputStream为String
-//	 *
-//	 * @param is
-//	 * @param encoding
-//	 * @return
-//	 */
-//	public static String inputStream2String(InputStream is, String encoding) {
-//		if (null == is) {
-//			return null;
-//		}
-//		StringBuilder buffer = new StringBuilder();
-//		BufferedReader in = null;
-//		try {
-//			if (StringUtil.isNotBlank(encoding)) {
-//				in = new BufferedReader(new InputStreamReader(is, encoding));
-//			} else {
-//				in = new BufferedReader(new InputStreamReader(is));
-//			}
-//			String line = "";
-//			while ((line = in.readLine()) != null) {
-//				buffer.append(line);
-//				buffer.append("\r\n");
-//			}
-//		} catch (Exception e) {
-//			e.printStackTrace();
-//			logger.error(e.getMessage());
-//		} finally {
-//			closeQuietly(in);
-//		}
-//		return buffer.toString();
-//	}
+public class MacroIfLogic {
 
 	/**
 	 * @todo 简单逻辑判断,只支持2个逻辑,update 2017-12-4 剔除freemarker复杂逻辑判断,减少框架依赖性
@@ -355,47 +257,5 @@ public class CommonUtils {
 			return Double.parseDouble(valueStr) < Double.parseDouble(compare);
 		}
 		return valueStr.compareTo(compare) < 0;
-	}
-
-	/**
-	 * @todo 加工字段名称，将数据库sql查询的columnName转成对应对象的属性名称(去除下划线)
-	 * @param labelNames
-	 * @return
-	 */
-	public static String[] humpFieldNames(String[] labelNames) {
-		if (labelNames == null)
-			return null;
-		String[] result = new String[labelNames.length];
-		for (int i = 0, n = labelNames.length; i < n; i++) {
-			result[i] = StringUtil.toHumpStr(labelNames[i], false);
-		}
-		return result;
-	}
-
-	/**
-	 * @todo 关闭一个或多个流对象
-	 * @param closeables 可关闭的流对象列表
-	 * @throws IOException
-	 */
-	public static void close(Closeable... closeables) throws IOException {
-		if (closeables != null) {
-			for (Closeable closeable : closeables) {
-				if (closeable != null) {
-					closeable.close();
-				}
-			}
-		}
-	}
-
-	/**
-	 * @todo 关闭一个或多个流对象
-	 * @param closeables 可关闭的流对象列表
-	 */
-	public static void closeQuietly(Closeable... closeables) {
-		try {
-			close(closeables);
-		} catch (IOException e) {
-			// do nothing
-		}
 	}
 }
